Make the registration point a standalone file
