+++ cmd/medius/publish.go
@@ -15,21 +15,11 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/ulikunitz/xz"
-	"kubevirt.io/containerdisks/artifacts/centos"
-	"kubevirt.io/containerdisks/artifacts/fedora"
-	"kubevirt.io/containerdisks/artifacts/rhcos"
 	"kubevirt.io/containerdisks/pkg/api"
 	"kubevirt.io/containerdisks/pkg/build"
 	"kubevirt.io/containerdisks/pkg/http"
 	"kubevirt.io/containerdisks/pkg/repository"
 )
-
-var registry = []api.Artifact{
-	fedora.New("35"),
-	rhcos.New("4.9"),
-	centos.New("8.4"),
-	centos.New("7-2009"),
-}
 
 type PublishOptions struct {
 	ForceBuild bool
